# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eBay Price Checker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface ResearchRequest {
  /** eBay item URL */
  url: string;
  /** Your current listing price (optional) */
  myPrice?: number;
}

export interface ProductIdentifiers {
  upc?: string[];
  ean?: string[];
  isbn?: string[];
  mpn?: string;
  brand?: string;
  epid?: string;
}

export interface EbayItem {
  itemId: string;
  title: string;
  price: number;
  currency: string;
  condition: string;
  conditionId?: string;
  seller?: string;
  url: string;
  shippingCost?: number;
  totalPrice: number;
  imageUrl?: string;
  location?: string;
  identifiers?: ProductIdentifiers;
}

export type ResearchResultLowestByCondition = { [key: string]: EbayItem };

export interface ResearchResult {
  originalItem: EbayItem;
  lowestByCondition: ResearchResultLowestByCondition;
  allItems: EbayItem[];
  evidenceUrls: string[];
  searchedAt: string;
}

export interface InventoryResearchRequest {
  inventoryProductId: number;
}

export interface InventoryProductSnapshot {
  id?: number;
  mercari_url?: string;
  ebay_url?: string;
  ebay_price_usd?: number;
  status?: string;
}

export type InventoryProductRow = InventoryProductSnapshot & {
  purchase_price?: number;
  alert_status?: string | null;
  last_check?: string | null;
  /** Existing monitor id if already tracked */
  monitorId?: number | null;
};

export interface InventoryProductsListResponse {
  count: number;
  products: InventoryProductRow[];
}

export type InventoryResearchResponse = ResearchResult & {
  inventoryProduct: InventoryProductSnapshot;
};

export interface CreateMonitorFromInventoryRequest {
  inventoryProductId: number;
  /** Optional. Pick a listing URL from research candidates; defaults to inventory ebay_url */
  seedEbayUrl?: string;
  /** Defaults to inventory ebay_price_usd when omitted */
  myPrice?: number;
  myCondition: string;
  /** Candidate condition selected by user */
  trackedTargetCondition?: string;
  /** Candidate current total price selected by user */
  trackedTargetPrice?: number;
  /** Write tracked target URL/condition/price to sheet immediately */
  syncToSpreadsheetNow?: boolean;
  label?: string;
  spreadsheetRow?: number;
  checkIntervalMinutes?: number;
}

export interface Monitor {
  id: number;
  ebayUrl: string;
  myPrice: number;
  myCondition: string;
  label?: string;
  spreadsheetRow?: number;
  inventoryProductId?: number;
  checkIntervalMinutes?: number;
  isActive: boolean;
  currentLowestPrice?: number;
  currentLowestCondition?: string;
  lastCheckedAt?: string;
  createdAt: string;
  hasAlert: boolean;
}

export type MonitorFromInventoryResponse = Monitor & {
  /** True when an existing monitor for this inventory row was updated */
  updated?: boolean;
  /** True when tracked target fields were written to sheet immediately */
  spreadsheetSynced?: boolean;
};

export interface CreateMonitorRequest {
  ebayUrl: string;
  myPrice: number;
  myCondition: string;
  label?: string;
  /** Row number in spreadsheet to update */
  spreadsheetRow?: number;
  inventoryProductId?: number;
  checkIntervalMinutes?: number;
}

export interface UpdateMonitorRequest {
  myPrice?: number;
  myCondition?: string;
  label?: string;
  spreadsheetRow?: number;
  inventoryProductId?: number | null;
  checkIntervalMinutes?: number;
  isActive?: boolean;
}

export interface PriceHistory {
  id: number;
  monitorId: number;
  lowestPrice: number;
  lowestCondition?: string;
  checkedAt: string;
}

export interface Alert {
  id: number;
  monitorId: number;
  monitorLabel?: string;
  competitorPrice: number;
  competitorCondition?: string;
  myPrice: number;
  priceDiff: number;
  isResolved: boolean;
  spreadsheetSynced: boolean;
  createdAt: string;
  resolvedAt?: string;
}

export type MonitorDetail = Monitor & {
  recentHistory?: PriceHistory[];
  activeAlerts?: Alert[];
};

export interface CheckResult {
  monitorId: number;
  lowestPrice: number;
  lowestCondition?: string;
  myPrice: number;
  alertTriggered: boolean;
  repricedTo?: number;
  evidenceUrls: string[];
  message?: string;
  checkedAt: string;
}

export interface SpreadsheetConfigRequest {
  spreadsheetId: string;
  sheetName: string;
  /** Column index (0-based) for price alert */
  alertColumnIndex?: number;
  /** Column index (0-based) for competitor price write-back */
  priceColumnIndex?: number;
  /** Column index (0-based) where my own price is stored in the sheet */
  myPriceColumnIndex?: number;
  sourceUrlColumnIndex?: number;
  ebayUrlColumnIndex?: number;
  lowestPriceColumnIndex?: number;
  lowestConditionColumnIndex?: number;
  lastCheckColumnIndex?: number;
  alertStatusColumnIndex?: number;
  repricedValueColumnIndex?: number;
  evidenceUrlsColumnIndex?: number;
  trackedTargetUrlColumnIndex?: number;
  trackedTargetConditionColumnIndex?: number;
  trackedTargetPriceColumnIndex?: number;
  autoRepriceEnabled?: boolean;
  undercutAmount?: number;
  minAllowedPrice?: number;
  ebayAppId?: string;
  ebayDevId?: string;
  ebayCertId?: string;
  ebayUserToken?: string;
  inventoryCheckerBaseUrl?: string;
  inventoryCheckerApiKey?: string;
  inventoryStatusColumnIndex?: number;
  ebayListingConditionColumnIndex?: number | null;
  ebayOAuthRedirectUri?: string;
  ebayOAuthClientSecret?: string;
  /** Google Service Account JSON (stored securely) */
  serviceAccountJson?: string;
}

export interface SpreadsheetConfig {
  spreadsheetId?: string;
  sheetName?: string;
  alertColumnIndex?: number;
  priceColumnIndex?: number;
  myPriceColumnIndex?: number;
  sourceUrlColumnIndex?: number;
  ebayUrlColumnIndex?: number;
  lowestPriceColumnIndex?: number;
  lowestConditionColumnIndex?: number;
  lastCheckColumnIndex?: number;
  alertStatusColumnIndex?: number;
  repricedValueColumnIndex?: number;
  evidenceUrlsColumnIndex?: number;
  trackedTargetUrlColumnIndex?: number;
  trackedTargetConditionColumnIndex?: number;
  trackedTargetPriceColumnIndex?: number;
  autoRepriceEnabled?: boolean;
  undercutAmount?: number;
  minAllowedPrice?: number;
  hasEbayCredentials?: boolean;
  hasEbayOAuth?: boolean;
  inventoryCheckerBaseUrl?: string;
  hasInventoryCheckerApiKey?: boolean;
  inventoryStatusColumnIndex?: number;
  ebayListingConditionColumnIndex?: number;
  ebayOAuthRedirectUri?: string;
  isConfigured: boolean;
  hasServiceAccount: boolean;
  /** True when Postgres schema is missing (run drizzle-kit push) */
  databaseNeedsMigration?: boolean;
  /** User-facing hint in Japanese */
  setupMessageJa?: string;
}

export interface SpreadsheetRowData {
  rowNumber: number;
  myPrice?: number;
  rawValue?: string;
  found: boolean;
}

export interface ConnectionTestResult {
  success: boolean;
  message: string;
  sheetTitle?: string;
}

export interface SyncResult {
  synced: number;
  failed: number;
  message: string;
}

export interface SheetSyncCounts {
  created: number;
  updated: number;
  skipped: number;
}

export interface InventorySyncResult {
  updated: number;
  created: number;
  skipped: boolean;
}

export interface SyncPipelineResult {
  sheet: SheetSyncCounts;
  inventory: InventorySyncResult;
  hydratedConditions: number;
  linkedRows: number;
}

export interface RemoteInventorySyncResult {
  [key: string]: unknown;
}

export interface EbayOAuthAuthorizeResponse {
  authorizeUrl: string;
}

export type EbayOAuthCallbackParams = {
  code: string;
};