# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: eBay Price Checker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /research:
    post:
      operationId: researchPrice
      tags: [research]
      summary: Research eBay item prices by URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchRequest"
      responses:
        "200":
          description: Price research results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/products:
    get:
      operationId: listInventoryProducts
      tags: [inventory]
      summary: Products from linked inventory service with monitor status
      responses:
        "200":
          description: Inventory product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryProductsListResponse"
        "400":
          description: Inventory URL not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/research:
    post:
      operationId: inventoryResearchProduct
      tags: [inventory]
      summary: Run same-item research for one inventory product listing URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryResearchRequest"
      responses:
        "200":
          description: Research results plus inventory snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResearchResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/monitors:
    post:
      operationId: createMonitorFromInventory
      tags: [inventory]
      summary: Create or update a monitor from inventory row
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMonitorFromInventoryRequest"
      responses:
        "200":
          description: Monitor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorFromInventoryResponse"
        "201":
          description: Monitor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorFromInventoryResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /monitors:
    get:
      operationId: listMonitors
      tags: [monitors]
      summary: List all price monitors
      responses:
        "200":
          description: List of monitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Monitor"
    post:
      operationId: createMonitor
      tags: [monitors]
      summary: Create a new price monitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMonitorRequest"
      responses:
        "201":
          description: Monitor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    ResearchRequest:
      type: object
      properties:
        url:
          type: string
          description: eBay item URL
        myPrice:
          type: number
          description: Your current listing price (optional)
      required:
        - url
    ProductIdentifiers:
      type: object
      properties:
        upc:
          type: array
          items:
            type: string
        ean:
          type: array
          items:
            type: string
        isbn:
          type: array
          items:
            type: string
        mpn:
          type: string
        brand:
          type: string
        epid:
          type: string
    EbayItem:
      type: object
      properties:
        itemId:
          type: string
        title:
          type: string
        price:
          type: number
        currency:
          type: string
        condition:
          type: string
        conditionId:
          type: string
        seller:
          type: string
        url:
          type: string
        shippingCost:
          type: number
        totalPrice:
          type: number
        imageUrl:
          type: string
        location:
          type: string
        identifiers:
          $ref: "#/components/schemas/ProductIdentifiers"
      required:
        - itemId
        - title
        - price
        - currency
        - condition
        - url
        - totalPrice
    ResearchResult:
      type: object
      properties:
        originalItem:
          $ref: "#/components/schemas/EbayItem"
        lowestByCondition:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/EbayItem"
        allItems:
          type: array
          items:
            $ref: "#/components/schemas/EbayItem"
        evidenceUrls:
          type: array
          items:
            type: string
        searchedAt:
          type: string
          format: date-time
      required:
        - originalItem
        - lowestByCondition
        - allItems
        - evidenceUrls
        - searchedAt
    InventoryResearchRequest:
      type: object
      required: [inventoryProductId]
      properties:
        inventoryProductId:
          type: integer
    InventoryProductSnapshot:
      type: object
      properties:
        id:
          type: integer
        mercari_url:
          type: string
        ebay_url:
          type: string
        ebay_price_usd:
          type: number
        status:
          type: string
    InventoryProductRow:
      allOf:
        - $ref: "#/components/schemas/InventoryProductSnapshot"
        - type: object
          properties:
            purchase_price:
              type: number
            alert_status:
              type: string
              nullable: true
            last_check:
              type: string
              nullable: true
            monitorId:
              type: integer
              nullable: true
              description: Existing monitor id if already tracked
    InventoryProductsListResponse:
      type: object
      required: [count, products]
      properties:
        count:
          type: integer
        products:
          type: array
          items:
            $ref: "#/components/schemas/InventoryProductRow"
    InventoryResearchResponse:
      allOf:
        - $ref: "#/components/schemas/ResearchResult"
        - type: object
          required: [inventoryProduct]
          properties:
            inventoryProduct:
              $ref: "#/components/schemas/InventoryProductSnapshot"
    CreateMonitorFromInventoryRequest:
      type: object
      required: [inventoryProductId, myCondition]
      properties:
        inventoryProductId:
          type: integer
        seedEbayUrl:
          type: string
          description: Optional. Pick a listing URL from research candidates; defaults to inventory ebay_url
        myPrice:
          type: number
          description: Defaults to inventory ebay_price_usd when omitted
        myCondition:
          type: string
        trackedTargetCondition:
          type: string
          description: Candidate condition selected by user
        trackedTargetPrice:
          type: number
          description: Candidate current total price selected by user
        syncToSpreadsheetNow:
          type: boolean
          description: Write tracked target URL/condition/price to sheet immediately
        label:
          type: string
        spreadsheetRow:
          type: integer
        checkIntervalMinutes:
          type: integer
          default: 360
    MonitorFromInventoryResponse:
      allOf:
        - $ref: "#/components/schemas/Monitor"
        - type: object
          properties:
            updated:
              type: boolean
              description: True when an existing monitor for this inventory row was updated
            spreadsheetSynced:
              type: boolean
              description: True when tracked target fields were written to sheet immediately
    CreateMonitorRequest:
      type: object
      properties:
        ebayUrl:
          type: string
        myPrice:
          type: number
        myCondition:
          type: string
        label:
          type: string
        spreadsheetRow:
          type: integer
          description: Row number in spreadsheet to update
        inventoryProductId:
          type: integer
        checkIntervalMinutes:
          type: integer
          default: 360
      required:
        - ebayUrl
        - myPrice
        - myCondition
...
    Monitor:
      type: object
      properties:
        id:
          type: integer
        ebayUrl:
          type: string
        myPrice:
          type: number
        myCondition:
          type: string
        label:
          type: string
        spreadsheetRow:
          type: integer
        inventoryProductId:
          type: integer
        checkIntervalMinutes:
          type: integer
        isActive:
          type: boolean
        currentLowestPrice:
          type: number
        currentLowestCondition:
          type: string
        lastCheckedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        hasAlert:
          type: boolean
      required:
        - id
        - ebayUrl
        - myPrice
        - myCondition
        - isActive
        - createdAt
        - hasAlert
...