# lib/api-zod/src/generated/types/createMonitorFromInventoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eBay Price Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMonitorFromInventoryRequest {
  inventoryProductId: number;
  /** Optional. Pick a listing URL from research candidates; defaults to inventory ebay_url */
  seedEbayUrl?: string;
  /** Defaults to inventory ebay_price_usd when omitted */
  myPrice?: number;
  myCondition: string;
  /** Candidate condition selected by user */
  trackedTargetCondition?: string;
  /** Candidate current total price selected by user */
  trackedTargetPrice?: number;
  /** Write tracked target URL/condition/price to sheet immediately */
  syncToSpreadsheetNow?: boolean;
  label?: string;
  spreadsheetRow?: number;
  checkIntervalMinutes?: number;
}