# lib/api-zod/src/generated/types/monitorFromInventoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eBay Price Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { Monitor } from "./monitor";

export type MonitorFromInventoryResponse = Monitor & {
  /** True when an existing monitor for this inventory row was updated */
  updated?: boolean;
  /** True when tracked target fields were written to sheet immediately */
  spreadsheetSynced?: boolean;
};